Move separator
