+++ src/console.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+namespace console {
+void print_separator();
+}
