+++ src/utils.hpp
@@ -5,7 +5,6 @@
 
 namespace utils {
 
-void print_separator();
 std::string read_from_file(const std::string &filename);
 void write_to_file(const std::string &filename, const std::string &text);
 std::filesystem::path get_project_data_dir();
